[BUA-1022] Renaming packages from tip to ep
